Changes to hopefully make it work

+++ enable-elastic-agent.sh
@@ -14,8 +14,10 @@ else
     echo lpass | sudo -S apt install unzip -y
 fi
 
-wget https://artifacts.elastic.co/downloads/beats/elastic-agent/elastic-agent-7.17.4-windows-x86_64.zip -O ansible/roles/windows/elastic-agent/files/elastic-agent-7.17.4-windows-x86_64.zip
-unzip ansible/roles/windows/elastic-agent/files/elastic-agent-7.17.4-windows-x86_64.zip -d ansible/roles/windows/elastic-agent/files/
+if [ ! -e ansible/roles/windows/elastic-agent/files/elastic-agent-7.17.4-windows-x86_64.zip]; then
+    wget https://artifacts.elastic.co/downloads/beats/elastic-agent/elastic-agent-7.17.4-windows-x86_64.zip -O ansible/roles/windows/elastic-agent/files/elastic-agent-7.17.4-windows-x86_64.zip
+    unzip ansible/roles/windows/elastic-agent/files/elastic-agent-7.17.4-windows-x86_64.zip -d ansible/roles/windows/elastic-agent/files/
+fi
 
 if [ ! -f umarker.txt ]
 then
